refactor: move update notification to module
